requirements.txt
Django==1.10.1
django-ckeditor==5.1.1
django-contrib-comments==1.7.3
django-crispy-forms==1.6.1
django-extensions==1.7.4
django-pure-pagination==0.3.0
django-taggit==0.21.2
Pillow==3.4.2
six==1.10.0


